pyproject.toml
[tool.poetry]
name = "ostrovok_tests"
version = "0.1.0"
description = "API, Mobile and UI testsOstrovok.ru"
authors = ["Sterphius <malyshff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
allure-pytest = "^2.13.0"
Appium-Python-Client = "^2.8.1"
curlify = "2.2.1"
pydantic = "^1.10.5"
pytest = "^7.1.3"
pytest-voluptuous = "^1.2.0"
pytest-xdist = "^2.5.0"
python = "^3.10.5"
python-dotenv = "^0.21.0"
requests = "^2.28.2"
selenium = "4.5.0"
selene = {allow-prereleases = true, version = "^2.0.0-beta.17"}
voluptuous = "^0.13.1"

[tool.pytest.ini_options]
addopts="-vv --alluredir=allure-results --clean-alluredir"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
